# src/main/java/com/sabrina/ecommerce/controller/api/AdminApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.ecommerce.controller.api;


import com.sabrina.ecommerce.controller.model.CategoryApiRequest;
import com.sabrina.ecommerce.controller.model.CategoryApiResponse;
import com.sabrina.ecommerce.controller.model.OrderApiResponse;
import com.sabrina.ecommerce.controller.model.ProductApiRequest;
import com.sabrina.ecommerce.controller.model.ProductApiResponse;
import com.sabrina.ecommerce.controller.model.UpdateOrderStatusRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-01T11:52:41.153654499+02:00[Europe/Andorra]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Admin", description = "the Admin API")
public interface AdminApi {

    /**
     * POST /api/admin/categories
     *
     * @param categoryApiRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "createCategory",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/admin/categories",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<CategoryApiResponse> createCategory(
        @Parameter(name = "CategoryApiRequest", description = "", required = true) @Valid @RequestBody CategoryApiRequest categoryApiRequest
    );


    /**
     * POST /api/admin/products
     *
     * @param productApiRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "createProduct",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/admin/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductApiResponse> createProduct(
        @Parameter(name = "ProductApiRequest", description = "", required = true) @Valid @RequestBody ProductApiRequest productApiRequest
    );


    /**
     * DELETE /api/admin/orders/{id}/status
     *
     * @param id  (required)
     * @return Deleted (status code 204)
     */
    @Operation(
        operationId = "deleteCategory",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/admin/orders/{id}/status"
    )
    
    ResponseEntity<Void> deleteCategory(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * DELETE /api/admin/products/{id}
     *
     * @param id  (required)
     * @return Deleted (status code 204)
     */
    @Operation(
        operationId = "deleteProduct",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Deleted")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/admin/products/{id}"
    )
    
    ResponseEntity<Void> deleteProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );


    /**
     * GET /api/admin/orders
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getAllOrders",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = OrderApiResponse.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/admin/orders",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<OrderApiResponse>> getAllOrders(
        
    );


    /**
     * PUT /api/admin/categories/{id}
     *
     * @param id  (required)
     * @param categoryApiRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateCategory",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/admin/categories/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<CategoryApiResponse> updateCategory(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "CategoryApiRequest", description = "", required = true) @Valid @RequestBody CategoryApiRequest categoryApiRequest
    );


    /**
     * PUT /api/admin/orders/{id}/status
     *
     * @param id  (required)
     * @param updateOrderStatusRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateOrderStatus",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = OrderApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/admin/orders/{id}/status",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<OrderApiResponse> updateOrderStatus(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "UpdateOrderStatusRequest", description = "", required = true) @Valid @RequestBody UpdateOrderStatusRequest updateOrderStatusRequest
    );


    /**
     * PUT /api/admin/products/{id}
     *
     * @param id  (required)
     * @param productApiRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateProduct",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductApiResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/admin/products/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProductApiResponse> updateProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "ProductApiRequest", description = "", required = true) @Valid @RequestBody ProductApiRequest productApiRequest
    );

}