# src/main/java/com/sabrina/ecommerce/controller/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.ecommerce.controller.api;

import com.sabrina.ecommerce.controller.model.AuthApiResponse;
import com.sabrina.ecommerce.controller.model.LoginApiRequest;
import com.sabrina.ecommerce.controller.model.RegisterApiRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-01T11:52:41.153654499+02:00[Europe/Andorra]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Auth", description = "the Auth API")
public interface AuthApi {

    /**
     * POST /api/auth/login
     *
     * @param loginApiRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "login",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthApiResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthApiResponse> login(
        @Parameter(name = "LoginApiRequest", description = "", required = true) @Valid @RequestBody LoginApiRequest loginApiRequest
    );


    /**
     * POST /api/auth/register
     *
     * @param registerApiRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "register",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/register",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> register(
        @Parameter(name = "RegisterApiRequest", description = "", required = true) @Valid @RequestBody RegisterApiRequest registerApiRequest
    );


    /**
     * POST /api/auth/register-admin
     *
     * @param registerApiRequest  (required)
     * @return Created (status code 201)
     */
    @Operation(
        operationId = "registerAdmin",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/auth/register-admin",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> registerAdmin(
        @Parameter(name = "RegisterApiRequest", description = "", required = true) @Valid @RequestBody RegisterApiRequest registerApiRequest
    );

}