# src/main/java/com/sabrina/ecommerce/controller/api/CategoryApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.ecommerce.controller.api;


import com.sabrina.ecommerce.controller.model.CategoryApiResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-01T11:52:41.153654499+02:00[Europe/Andorra]", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Category", description = "the Category API")
public interface CategoryApi {

    /**
     * GET /api/categories
     *
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getAllCategories",
        tags = { "Category" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CategoryApiResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/categories",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<CategoryApiResponse>> getAllCategories(
        
    );


    /**
     * GET /api/categories/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCategoryById",
        tags = { "Category" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryApiResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/categories/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<CategoryApiResponse> getCategoryById(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    );

}