# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  version: 1.0.0
  description: API First for ecommerce project
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: [ ]
...
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterApiRequest"
      responses:
        "201":
          description: Created
  /api/auth/register-admin:
    post:
      tags:
        - Auth
      operationId: registerAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterApiRequest"
      responses:
        "201":
          description: Created
  /api/auth/login:
    post:
      tags:
        - Auth
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginApiRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthApiResponse"
...
  /api/admin/products:
    post:
        tags:
          - Admin
        operationId: createProduct
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiRequest"
        responses:
          "201":
            description: Created
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ProductApiResponse"
  /api/admin/products/{id}:
    put:
        tags:
          - Admin
        operationId: updateProduct
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductApiRequest"
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ProductApiResponse"

    delete:
        tags:
          - Admin
        operationId: deleteProduct
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          "204":
            description: Deleted
...
  /api/categories/{id}:
   get:
      tags:
        - Category
      operationId: getCategoryById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiResponse"
  /api/admin/categories:
      post:
        tags:
          - Admin
        operationId: createCategory
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiRequest"
        responses:
          "201":
            description: Created
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CategoryApiResponse"
  /api/admin/categories/{id}:
      put:
        tags:
          - Admin
        operationId: updateCategory
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryApiRequest"
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CategoryApiResponse"
...
  /api/admin/orders/{id}/status:
    put:
        tags:
          - Admin
        operationId: updateOrderStatus
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        responses:
          "200":
            description: OK
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/OrderApiResponse"

    delete:
        tags:
          - Admin
        operationId: deleteCategory
        parameters:
          - name: id
            in: path
            required: true
            schema:
              type: integer
        responses:
          "204":
            description: Deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginApiRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    RegisterApiRequest:
      type: object
      required:
        - email
        - password
        - username
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
    AuthApiResponse:
      type: object
      properties:
        token:
          type: string
        email:
          type: string
        role:
          type: string
    ProductApiResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
    ProductApiRequest:
      type: object
      required:
        - name
        - price
        - categoryId
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        categoryId:
          type: integer
...
    CategoryApiRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
    CategoryApiResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
    OrderApiResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        totalPrice:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemApiResponse"
...
    OrderItemApiResponse:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
          format: double